requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
distro==1.4.0
Django==2.1.5
django-bulk-update==2.2.0
django-celery==3.2.2
django-redis==4.10.0
django-stubs==0.12.1
djangorestframework==3.9.1
djangorestframework-stubs==0.2.0
kombu==3.0.37
mypy==0.701
mypy-extensions==0.4.1
mysqlclient==1.4.1
Pillow==5.4.1
pytz==2018.9
redis==2.10.6
shortuuid==0.5.0
sqlparse==0.2.4
typed-ast==1.3.4
typing-extensions==3.7.2
xlrd==1.2.0
